ecommerce_project/requirements.txt
aiosmtplib==2.0.2
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.1.2
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.1.1
fastapi==0.110.0
fastapi-mail==1.4.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
PyJWT==2.10.1
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
requests==2.31.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.36.3
stripe==12.0.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.29.0
slowapi==0.1.9
httpx==0.24.0
google-auth==3.0.0
